# gen/models/protected_entity_snapshot_id.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"
)

// ProtectedEntitySnapshotID protected entity snapshot ID
// swagger:model ProtectedEntitySnapshotID
type ProtectedEntitySnapshotID string

// Validate validates this protected entity snapshot ID
func (m ProtectedEntitySnapshotID) Validate(formats strfmt.Registry) error {
	return nil
}