# gen/models/task_nexus_id.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"
)

// TaskNexusID task nexus ID
// swagger:model TaskNexusID
type TaskNexusID string

// Validate validates this task nexus ID
func (m TaskNexusID) Validate(formats strfmt.Registry) error {
	return nil
}